refactor: extract post list action buttons
